Added reasonable defaults for HTTP header caching.
These make serving long-running aggregations as end-points for the faceted browser feel much faster after initial load.

From the FAQ:

Because the opening page of a faceted search is also the most computationally-intensive to produce (it integrates
counts across the entire dataset), the faceted browser is radically faster when configured to cache results. As of version
0.7.4, the controller mixin sets HTTP cache headers in responses to reasonable defaults.

The short story is that models should have an updated_at column, which is used to compute cache keys. For models with many
items, you may wish to add an index on updated_at. ...

(see FAQ for complete information)

Released as version 0.7.4